Reorganize session.* and tests
