+++ borderpatrol-core/src/main/scala/com/lookout/borderpatrol/session/package.scala
@@ -8,7 +8,6 @@ import org.jboss.netty.handler.codec.http.HttpRequest
 import scala.util.Try
 
 package object session {
-
   object Constants {
     object SessionId {
       val entropySize = 16
@@ -20,6 +19,19 @@ package object session {
       val lifetime = Duration(1, TimeUnit.DAYS)
     }
   }
+
+  sealed trait Session {
+    val id: SessionId
+    val originalRequest: HttpRequest
+    val tokens: Tokens
+  }
+
+  case class NewSession(originalRequest: HttpRequest)(implicit g: SessionIdGenerator, s: SecretStoreApi) extends Session {
+    lazy val id = g.next
+    val tokens = Tokens(EmptyToken, EmptyServiceTokens)
+  }
+
+  case class ExistingSession(id: SessionId, originalRequest: HttpRequest, tokens: Tokens) extends Session
 
   implicit class SessionIdSerialize(val s: SessionId) extends AnyVal {
     def asString(implicit marshaller: SessionIdMarshaller): String =
@@ -42,10 +54,26 @@ package object session {
     implicit val marshaller = SessionIdMarshaller(secretStore)
     implicit val generator: SessionIdGenerator = new SessionIdGenerator
     val sessionStore = new InMemorySessionStore
+  }
 
-    def apply(request: HttpRequest): Session = NewSession(request)
+  object Session {
+    def sessionStore: SessionStoreApi = SecureSession.sessionStore
+    implicit def generator: SessionIdGenerator = SecureSession.generator
+    implicit def secretStore: SecretStoreApi = SecureSession.secretStore
 
-    def apply(s: String, request: HttpRequest): Session =
-      sessionStore.get(s) getOrElse NewSession(request)
+    def apply(id: SessionId, originalRequest: HttpRequest, tokens: Tokens): Session =
+      save(ExistingSession(id, originalRequest, tokens))
+
+    def apply(s: String, originalRequest: HttpRequest): Session =
+      sessionStore.get(s) getOrElse newSession(originalRequest)
+
+    def apply(request: RoutedRequest): Session =
+      request.borderCookie.flatMap(id => sessionStore.get(id)) getOrElse newSession(request.httpRequest)
+
+    def newSession(originalRequest: HttpRequest): Session =
+      save(NewSession(originalRequest))
+
+    def save(session: Session): Session =
+      sessionStore.update(session)
   }
 }
