+++ borderpatrol-core/src/test/scala/com/lookout/borderpatrol/session/SessionSpec.scala
@@ -1,0 +1,15 @@
+package com.lookout.borderpatrol.session
+
+import org.scalatest.{Matchers, FlatSpec}
+
+class SessionSpec extends FlatSpec with Matchers {
+  implicit val secretStore = InMemorySecretStore(Secrets(Current(SecretExpiry.currentExpiry), None))
+  implicit val marshaller = SessionIdMarshaller(secretStore)
+  implicit val generator: SessionIdGenerator = new SessionIdGenerator
+  def mockSessionStore = new InMemorySessionStore
+
+  behavior of "Session"
+
+  it should "create and save a new session" in {
+  }
+}
